improve log messages, begin reformating of stepper driver (const getters, private methods)

+++ niryo_robot_hardware_stack/dynamixel_driver/src/synchronize_motor_cmd.cpp
@@ -79,7 +79,7 @@ namespace DynamixelDriver
         else {
             ss << "(";
             for(int i = 0; i < _motor_id_list.size(); ++i)
-                ss << " motor " << _motor_id_list.at(i) << ": " << _param_list.at(i) << ",";
+                ss << " motor " << static_cast<int>(_motor_id_list.at(i)) << ": " << _param_list.at(i) << ",";
 
         }
